library became executable

+++ basicstat.h
@@ -13,23 +13,23 @@ namespace snu {
 		double reciprocity; // only defined in directed graph
 		double negativity;
 		unsigned int lcc; // largest connected component
-		unsigned int four_tour_count;
+		unsigned long long four_tour_count;
 		unsigned int diameter; // diameter of graph
 	};
 
 	struct USResult {
-		int size;
-		int volume;
+		unsigned int size;
+		unsigned int volume;
 		double avg_degree; // average degree, 2m / n
 		double fill; // 2m / (N * (N + 1)), undirected graph with loops
-		int max_degree; // maximum degree
+		unsigned int max_degree; // maximum degree
 		double negativity;
 		unsigned int lcc;
-		unsigned int wedge_count;
-		unsigned int claw_count;
-		unsigned int triangle_count;
-		unsigned int square_count;
-		unsigned int four_tour_count;
+		unsigned long long wedge_count;
+		unsigned long long claw_count;
+		unsigned long long triangle_count;
+		unsigned long long square_count;
+		unsigned long long four_tour_count;
 		unsigned int diameter;
 	};
